various changes (= dj/220) (#318)
ENCLONE VISUAL
- Make the text size for the main part of the summary larger, thus increasing readability.
- Update enclone visual cookbook by recomputing it.
- Make enclone visual tooltip box position configurable and improve its appearance.
- Add two metrics to display.
OTHER
- Fix bugs in data import.
- Start building formal variable registry.  This will take a while.
- Add ways to get more info for top feature barcodes.
- Add feature barcode alluvial tables (internal for now).
- Import denovo stuff.
- cargo clippy --fix --all-targets --allow-dirty --workspace --exclude enclone_visual --exclude enclone_exec
STATUS
- ./test, enclone.test and test_vis pass

+++ enclone_tail/src/neighbor.rs
@@ -69,7 +69,7 @@ pub fn neighbor_joining(d: &Vec<Vec<f64>>) -> Vec<(usize, usize, f64)> {
         edges[2 * (n0 - n)] = (verts[f], vnew, df);
         edges[2 * (n0 - n) + 1] = (verts[g], vnew, dg);
         verts[f] = vnew;
-        for k in g..n - 1 as usize {
+        for k in g..n - 1_usize {
             verts[k] = verts[k + 1];
         }
         for i in 0..n {
